Separate test utils from tests for utils
