+++ src/test/java/uk/gov/hmcts/reform/blobrouter/testutils/SigningHelper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.blobrouter.util;
+package uk.gov.hmcts.reform.blobrouter.testutils;
 
 import java.security.KeyFactory;
 import java.security.Signature;
@@ -6,7 +6,7 @@ import java.security.spec.PKCS8EncodedKeySpec;
 
 final class SigningHelper {
 
-    static byte[] signWithSha256Rsa(byte[] input, byte[] keyBytes) throws Exception {
+    public static byte[] signWithSha256Rsa(byte[] input, byte[] keyBytes) throws Exception {
 
         Signature signature = Signature.getInstance("SHA256withRSA");
         signature.initSign(KeyFactory.getInstance("RSA").generatePrivate(new PKCS8EncodedKeySpec(keyBytes)));
